improve postinstall script removal

+++ script/transform-package/restore-postinstall.js
@@ -1,8 +1,10 @@
-const { readFileSync, writeFileSync, unlinkSync } = require("fs")
+const { existsSync, readFileSync, writeFileSync, unlinkSync } = require("fs")
 const { resolve } = require("path")
 
 const PACKAGE_PATH = resolve(__dirname, "../../package.json")
 const PACKAGE_CACHED_PATH = resolve(__dirname, "../../cached-package.json")
 
-writeFileSync(PACKAGE_PATH, readFileSync(PACKAGE_CACHED_PATH))
-unlinkSync(PACKAGE_CACHED_PATH)
+if (existsSync(PACKAGE_CACHED_PATH)) {
+  writeFileSync(PACKAGE_PATH, readFileSync(PACKAGE_CACHED_PATH))
+  unlinkSync(PACKAGE_CACHED_PATH)
+}
